Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    int a;
+     long int a;
     scanf("%d",&a);
     if(a&0){
         printf("Set");
